Adapt UIStorage to external sessions
